Code for aggregating scores on genes

+++ motifs/scripts/qsub/aggregateMotifScores.sh
@@ -27,13 +27,15 @@ OPTIONS:
    -g FILE File mapping states to genes.
    -s STR  Suffix of output files.
    -f STR  Extra suffix for output gene files.
+   -n      Create off matrices with sequences NOT overlapping the input beds.
 EOF
 }
 
 SUF=""
 SUF2=""
 GENEMAP=
-while getopts "hs:g:f:" opt
+MAKEOFF=0
+while getopts "hs:g:f:n" opt
 do
     case $opt in
 	h)
@@ -44,6 +46,8 @@ do
 	    SUF2=$OPTARG;;
 	g)
 	    GENEMAP=$OPTARG;;
+	n)
+	    MAKEOFF=1;;
 	?)
 	    usage
             exit 1;;
@@ -73,27 +77,51 @@ while read -r bedfile; do
     pref=${pref/.bed/}
     outfile1=${OUTDIR}/${pref}${SUF}_scores.npz
     outfile2=${GENEDIR}/${pref}${SUF}${SUF2}_gene_scores.npz
+    outfile1off=${OUTDIR}/${pref}${SUF}_OFF_scores.npz
+    outfile2off=${GENEDIR}/${pref}${SUF}${SUF2}_OFF_gene_scores.npz
+
     mapfile=${OUTDIR}/tmp/${pref}${SUF}_state_map.txt
     script=${OUTDIR}/tmp/${pref}${SUF}.sh
     errfile=${OUTDIR}/tmp/${pref}${SUF}.err
 
-    if [[ -z $GENEMAP ]] && [[ -f $outfile1 ]]; then
-	continue
+    if [[ -z $GENEMAP ]] && [[ -f $outfile1 ]]; then 
+	if [[ $MAKEOFF -eq  0 ]]; then
+	    continue
+	fi
+	if [[ $MAKEOFF -eq 1 ]] && [[ -f $outfile1off ]]; then
+	    continue
+	fi
     fi
     if [[ ! -z $GENEMAP ]] && [[ -f $outfile2 ]]; then
-	continue
+	if [[ $MAKEOFF -eq  0 ]]; then
+	    continue
+	fi 
+	if [[ $MAKEOFF -eq 1 ]] && [[ -f $outfile2off ]]; then
+	    continue
+	fi
     fi
 
     echo "#!/bin/bash" > $script
     echo "module add python/2.7" >> $script
     echo "module add bedtools/2.18.0" >> $script
-    echo "intersectBed -a $STATEMAP -b $bedfile -wa -u | awk 'BEGIN{OFS=\"\t\"}{print \$4,\$5}' > $mapfile" >> $script
     if [ ! -f $outfile1 ]; then
+	echo "intersectBed -a $STATEMAP -b $bedfile -wa -u | awk 'BEGIN{OFS=\"\t\"}{print \$4,\$5}' | sort -u > $mapfile" >> $script
 	echo "ls ${SCANDIR}/*npz | python $SRCDIR/python/combineRegionScoresMultiFile.py $mapfile $outfile1" >> $script
     fi
-    if [ ! -z $GENEMAP ]; then
+    if [[ ! -z $GENEMAP ]] && [[ ! -f $outfile2 ]] ; then
 	echo "ls $outfile1 | python $SRCDIR/python/combineRegionScoresMultiFile.py $GENEMAP $outfile2 --cmax" >> $script
     fi
+
+    if [[ $MAKEOFF -eq 1 ]]; then
+	if [ ! -f $outfile1off ]; then
+	    echo "intersectBed -a $STATEMAP -b $bedfile -wa -v | awk 'BEGIN{OFS=\"\t\"}{print \$4,\$5}' | sort -u > $mapfile" >> $script
+	    echo "ls ${SCANDIR}/*npz | python $SRCDIR/python/combineRegionScoresMultiFile.py $mapfile $outfile1off" >> $script
+	fi
+	if [[ ! -z $GENEMAP ]] && [[ ! -f $outfile2off ]] ; then
+	    echo "ls $outfile1off | python $SRCDIR/python/combineRegionScoresMultiFile.py $GENEMAP $outfile2off --cmax" >> $script
+	fi
+    fi
+
     qsub -N $pref -P large_mem -l h_vmem=48G -l h_rt=6:00:00 -e $errfile -o /dev/null $script
 done
